Fix: Prevent buffer overflow
Follow Tabnine and CodeQL's guidance and utilizes cin.width before the user input to prevent a buffer overflow.  Only the first 19 characters are read from the input buffer and the rest is ignored. Add a const buffer limit used for the array and width for easier maintainability if size needs to change. 

 Resolves: issue #1